feat: Add All Charts (Including Bivariate and Multivariate)
In this commit, I finished the Univariate analysis, including
all the observations and final section comments.

I also added all the charts from the Bivariate analysis, including
all the observations. The final section comments are pending, however.

Finally, from the Multivariate analysis, I added all the plots.
The observations from the second to last visualization is pending,
as are the final section comments.

On a later commit, I'll also add the Conclusion and finally finish
Part I of the project.