docs: Improve Key Insights Section
Initially, I added the key insights in the present tense, but now
that I've finished the second part (next commit) of the project, I
decided to change section to the past tense.

I also updated key insights to provide a better explanation as to why
I've decided to include those three charts in my explanatory
presentation.